+++ config/define.js
@@ -1,2 +1,7 @@
 // トップに出すPGタグの数
-const TOP_VIEW_TOTAL = 10;
+const TOP_VIEW_TOTAL_NUMBER = 10;
+
+module.exports = {
+    TOP_VIEW_TOTAL_NUMBER: TOP_VIEW_TOTAL_NUMBER
+};
+
